feat(front): adds position server route and page (#976)
Closes: #976
