+++ front/src/routes/cv/position/+page.js
@@ -1,0 +1,4 @@
+import { redirect } from '@sveltejs/kit';
+export function load() {
+  throw redirect(301, '/cv/positions');
+}
