+++ front/src/routes/cv/position/[slug]/+page.svelte
@@ -8,19 +8,21 @@
   // TODO: Import content mappers from a helpers library
   // then use the collection slug to select the api and map content appropriately
 
-  const { pathname, page, pageMeta } = data;
-  const { title, hero, intro } = page;
+  const { pathname, position, pageMeta } = data;
+  const { title, hero, body } = position;
 </script>
 
 <PageTransition transitionKey={pathname}>
-  <section class="cv">
+  <section class="cv meat">
     <Flourish />
     <a id="main">Main Content</a>
-    <h1>{title}</h1>
     <img src={hero.url} alt={hero.alternativeText} />
-    <div class="summary">
-      {intro}
-    </div>
+    <h1>{title}</h1>
+    {#if body}
+      <article class="position-longform">
+        {body}
+      </article>
+    {/if}
   </section>
 </PageTransition>
 <Meta {pageMeta} />
